+++ src/main/java/insane96mcp/progressivebosses/module/dragon/feature/HealthFeature.java
@@ -24,11 +24,13 @@ public class HealthFeature extends Feature {
 	private final ForgeConfigSpec.ConfigValue<Double> bonusRegenConfig;
 	private final ForgeConfigSpec.ConfigValue<Double> maximumBonusRegenConfig;
 	private final ForgeConfigSpec.ConfigValue<Double> bonusCrystalRegenConfig;
+	private final ForgeConfigSpec.DoubleValue bonusRegenRatioWhenHitConfig;
 
 	public double bonusHealth = 200d;
 	public double bonusRegen = 1.0d;
 	public double maxBonusRegen = 1.0d;
 	public double bonusCrystalRegen = 0d;
+	public double bonusRegenRatioWhenHit = 0.4d;
 
 	public HealthFeature(Module module) {
 		super(Config.builder, module);
@@ -45,6 +47,9 @@ public class HealthFeature extends Feature {
 		this.bonusCrystalRegenConfig = Config.builder
 				.comment("How much health (when missing 100% health) will the Ender Dragon regen at max difficulty each second whenever she's attached to a Crystal. So if she's missing 30% health, this will be 30% effective. This is added to the normal Crystal regen.")
 				.defineInRange("Bonus Crystal Regeneration", this.bonusCrystalRegen, 0.0, Double.MAX_VALUE);
+		this.bonusRegenRatioWhenHitConfig = Config.builder
+				.comment("Bonus regeneration (also crystal bonus regeneration) will be multiplied by this ratio when the Dragon has been hit in the last 3 seconds.")
+				.defineInRange("Bonus Regeneration Ratio When Hit", bonusRegenRatioWhenHit, 0.0d, 2.0d);
 		Config.builder.pop();
 	}
 
@@ -55,6 +60,7 @@ public class HealthFeature extends Feature {
 		this.maxBonusRegen = this.maximumBonusRegenConfig.get();
 		this.bonusRegen = this.bonusRegenConfig.get();
 		this.bonusCrystalRegen = this.bonusCrystalRegenConfig.get();
+		this.bonusRegenRatioWhenHit = this.bonusRegenRatioWhenHitConfig.get();
 	}
 
 	@SubscribeEvent
@@ -93,6 +99,9 @@ public class HealthFeature extends Feature {
 			return;
 
 		heal /= 20f;
+
+		if (dragon.tickCount - dragon.getLastHurtByMobTimestamp() <= 60) // 3 seconds
+			heal *= this.bonusRegenRatioWhenHit;
 
 		dragon.heal(heal);
 	}