+++ src/main/java/insane96mcp/progressivebosses/module/wither/feature/HealthFeature.java
@@ -21,10 +21,12 @@ public class HealthFeature extends Feature {
 	private final ForgeConfigSpec.ConfigValue<Double> bonusHealthConfig;
 	private final ForgeConfigSpec.ConfigValue<Double> maximumBonusRegenConfig;
 	private final ForgeConfigSpec.ConfigValue<Double> bonusRegenConfig;
+	private final ForgeConfigSpec.DoubleValue bonusRegenRatioWhenHitConfig;
 
 	public double bonusHealth = 720d;
 	public double maxBonusRegen = 2d;
 	public double bonusRegen = 2.4d;
+	public double bonusRegenRatioWhenHit = 0.4d;
 
 	public HealthFeature(Module module) {
 		super(Config.builder, module);
@@ -41,6 +43,9 @@ public class HealthFeature extends Feature {
 		this.bonusRegenConfig = Config.builder
 				.comment("How many half hearts will the Wither regen at max difficulty. This is added to the natural regeneration of the Wither (1 Health per Second).")
 				.defineInRange("Bonus Regeneration", bonusRegen, 0.0, Double.MAX_VALUE);
+		this.bonusRegenRatioWhenHitConfig = Config.builder
+				.comment("Bonus regeneration will be multiplied by this ratio when the Wither has been hit in the last 3 seconds.")
+				.defineInRange("Bonus Regeneration Ratio When Hit", bonusRegenRatioWhenHit, 0.0d, 2.0d);
 		Config.builder.pop();
 	}
 
@@ -50,6 +55,7 @@ public class HealthFeature extends Feature {
 		this.bonusHealth = this.bonusHealthConfig.get();
 		this.maxBonusRegen = this.maximumBonusRegenConfig.get();
 		this.bonusRegen = this.bonusRegenConfig.get();
+		this.bonusRegenRatioWhenHit = this.bonusRegenRatioWhenHitConfig.get();
 	}
 
 	@SubscribeEvent
@@ -81,6 +87,9 @@ public class HealthFeature extends Feature {
 
 		float heal = (float) Math.min(this.bonusRegen * DifficultyHelper.getScalingDifficulty(wither), this.maxBonusRegen);
 		heal /= 20f;
+
+		if (wither.tickCount - wither.getLastHurtByMobTimestamp() <= 60) // 3 seconds
+			heal *= this.bonusRegenRatioWhenHit;
 
 		if (heal > 0f)
 			wither.heal(heal);